Print method for objects of class factorplot

Prints the output from an object of class factorplot. By default, the function prints all pairwise differences along with standard errors and p-values (optionally adjusted for multiple testing). Optionally, it can print only significant differences.

# S3 method for factorplot
print(x, ..., digits = 3, sig = FALSE,
  trans = NULL)

Arguments

x
An object of class factorplot.

...
Other arguments passed to print, currently not implemented

digits
The number of digits to print in each column

sig
Logical indicating whether only significant differences should be printed.

trans
A character string representing the post-hypothesis-testing transformation to be performed on the estimates. For example, if the estimates provided to the factorplot command are log-floating absolute risks, you could use the transformation ‘exp’. The transformation is performed through a call to do.call

See also

Examples

est1 <- log(c(1.00,2.12,1.44,1.31,1.44,
     1.46,0.90))
var1 <- c(0.242,0.096,0.156,0.140,
     0.380,0.484,0.375)^2
names(est1) <- c(
     "Normal & superficial gastritis",
     "Chronic gastritis",
     "Chronic atrophic gastritits",
     "Intestinal metaplasia I",
     "Intestinal metaplasia II",
     "Intestinal metaplasia III",
     "Dysplasia")
plummer_fp1 <- factorplot(est1, var=var1, resdf=Inf)
print(plummer_fp1, trans="exp")

#>                                                                 Difference
#> Normal_&_superficial_gastritis -              Chronic_gastritis      0.472
#> Normal_&_superficial_gastritis -    Chronic_atrophic_gastritits      0.694
#>              Chronic_gastritis -    Chronic_atrophic_gastritits      1.472
#> Normal_&_superficial_gastritis -        Intestinal_metaplasia_I      0.763
#>              Chronic_gastritis -        Intestinal_metaplasia_I      1.618
#>    Chronic_atrophic_gastritits -        Intestinal_metaplasia_I      1.099
#> Normal_&_superficial_gastritis -       Intestinal_metaplasia_II      0.694
#>              Chronic_gastritis -       Intestinal_metaplasia_II      1.472
#>    Chronic_atrophic_gastritits -       Intestinal_metaplasia_II      1.000
#>        Intestinal_metaplasia_I -       Intestinal_metaplasia_II      0.910
#> Normal_&_superficial_gastritis -      Intestinal_metaplasia_III      0.685
#>              Chronic_gastritis -      Intestinal_metaplasia_III      1.452
#>    Chronic_atrophic_gastritits -      Intestinal_metaplasia_III      0.986
#>        Intestinal_metaplasia_I -      Intestinal_metaplasia_III      0.897
#>       Intestinal_metaplasia_II -      Intestinal_metaplasia_III      0.986
#> Normal_&_superficial_gastritis -                      Dysplasia      1.111
#>              Chronic_gastritis -                      Dysplasia      2.356
#>    Chronic_atrophic_gastritits -                      Dysplasia      1.600
#>        Intestinal_metaplasia_I -                      Dysplasia      1.456
#>       Intestinal_metaplasia_II -                      Dysplasia      1.600
#>      Intestinal_metaplasia_III -                      Dysplasia      1.622
#>                                                                    SE p.val
#> Normal_&_superficial_gastritis -              Chronic_gastritis 0.260 0.004
#> Normal_&_superficial_gastritis -    Chronic_atrophic_gastritits 0.288 0.205
#>              Chronic_gastritis -    Chronic_atrophic_gastritits 0.183 0.035
#> Normal_&_superficial_gastritis -        Intestinal_metaplasia_I 0.280 0.334
#>              Chronic_gastritis -        Intestinal_metaplasia_I 0.170 0.005
#>    Chronic_atrophic_gastritits -        Intestinal_metaplasia_I 0.210 0.652
#> Normal_&_superficial_gastritis -       Intestinal_metaplasia_II 0.451 0.418
#>              Chronic_gastritis -       Intestinal_metaplasia_II 0.392 0.324
#>    Chronic_atrophic_gastritits -       Intestinal_metaplasia_II 0.411 1.000
#>        Intestinal_metaplasia_I -       Intestinal_metaplasia_II 0.405 0.815
#> Normal_&_superficial_gastritis -      Intestinal_metaplasia_III 0.541 0.484
#>              Chronic_gastritis -      Intestinal_metaplasia_III 0.493 0.450
#>    Chronic_atrophic_gastritits -      Intestinal_metaplasia_III 0.509 0.978
#>        Intestinal_metaplasia_I -      Intestinal_metaplasia_III 0.504 0.830
#>       Intestinal_metaplasia_II -      Intestinal_metaplasia_III 0.615 0.982
#> Normal_&_superficial_gastritis -                      Dysplasia 0.446 0.813
#>              Chronic_gastritis -                      Dysplasia 0.387 0.027
#>    Chronic_atrophic_gastritits -                      Dysplasia 0.406 0.247
#>        Intestinal_metaplasia_I -                      Dysplasia 0.400 0.348
#>       Intestinal_metaplasia_II -                      Dysplasia 0.534 0.379
#>      Intestinal_metaplasia_III -                      Dysplasia 0.612 0.429